Move clientconfig functions to functions.inc.php
