+++ data/web/mobileconfig.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'inc/clientconfig.inc.php';
 require_once 'inc/prerequisites.inc.php';
 
 if (empty($mailcow_hostname)) {
